Search Results Page
This commit contains the design for the results page where the user will land after using the search functionality from the Main Page. It also includes a preview version of the Forum's design and a file that will serve as a guide for the rest of the team to use the same Navbar and Footer design across the whole website.