Landing Page Recommendations algorithm
Algorithm for book recommendations on the Landing Page has been implemented

If an user is logged in, books are recommended by the author and genres such user has read the most books from.

If the user is not logged in, books from the most read authors and genres appear as recommended

Custom texts added depending on whether the user has logged in or not.

Pending: correct pagination for the AJAX call.