Book Service Implemented
In order to guide the rest of the team on how we should implement the services, the Book's service has been implemented and tested over the REST API.

Models should be created for each kind of resource returned from the REST API in order to access each of their fields. The model MUST match the JSON's structure returned by the REST API.

Favicon has been changed to Bookmark's logo.

Fixed Proxy's Configuration file routing rule (had a single * when it should have two **).