Updated Navbar and Profile page
Navbar now knows when to show the profile Dropdown, and the Dropdown has been implemented.

Navbar is now fully functional.

The profile page now shows info of the user that's specified in the URL and checks if it is the current logged user in order to show the modify profile button or not.

Profile page is now fully functional.